requirements.txt
asgiref==3.3.1
boto3==1.17.12
botocore==1.20.12
cffi==1.14.5
cryptography==3.4.6
Django==3.1.7
django-filter==2.4.0
django-rest-knox==4.1.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
importlib-metadata==3.4.0
jmespath==0.10.0
Markdown==3.3.3
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.0
